+++ mvi/sample/src/main/java/ru/surfstudio/android/core/mvi/sample/ui/screen/main/MainMiddleware.kt
@@ -1,17 +1,18 @@
 package ru.surfstudio.android.core.mvi.sample.ui.screen.main
 
 import io.reactivex.Observable
+import ru.surfstudio.android.core.mvi.impls.ui.dialog.standard.CustomAlertDialogRoute
 import ru.surfstudio.android.core.mvi.impls.ui.middleware.BaseMiddleware
 import ru.surfstudio.android.core.mvi.impls.ui.middleware.BaseMiddlewareDependency
 import ru.surfstudio.android.core.mvi.impls.ui.middleware.navigation.composition.NavigationMiddleware
-import ru.surfstudio.android.core.mvi.impls.ui.middleware.navigation.composition.ObserveResult
-import ru.surfstudio.android.core.mvi.impls.ui.middleware.navigation.composition.listenForResult
-import ru.surfstudio.android.core.mvi.impls.ui.middleware.navigation.composition.observeResult
+import ru.surfstudio.android.core.mvi.impls.ui.middleware.navigation.composition.listenForResultMap
+import ru.surfstudio.android.core.mvi.impls.ui.middleware.navigation.composition.open
 import ru.surfstudio.android.core.mvi.sample.ui.screen.input.InputFormActivityRoute
+import ru.surfstudio.android.core.mvi.sample.ui.screen.main.MainEvent.Navigation
 import ru.surfstudio.android.core.ui.navigation.ScreenResult
 import ru.surfstudio.android.dagger.scope.PerScreen
 import ru.surfstudio.android.message.MessageController
-import ru.surfstudio.android.core.mvi.sample.ui.screen.main.MainEvent.*
+import ru.surfstudio.android.rx.extension.toObservable
 import javax.inject.Inject
 
 /**
@@ -28,18 +29,37 @@ class MainMiddleware @Inject constructor(
         addAll(
                 Navigation::class decomposeTo navigationMiddleware,
 //        Когда нужна простая реакция на событие - можно использовать listenForResult с колбеком
-                listenForResult(InputFormActivityRoute::class.java, ::Navigation, ::showResult)
+//              listenForResult(InputFormActivityRoute::class.java, ::Navigation, ::showResultReact)
 //        Когда нужна трансформация результата в событие - используется listenForResultMap
-//        +listenForResultMap(InputFormActivityRoute::class.java) {
-//            showResult(it)
-//            Navigation().open(ComplexListActivityRoute()).toObservable())
+                listenForResultMap(InputFormActivityRoute::class.java, ::Navigation, ::showResultMap)
 //        }
         )
 
     }
 
-    private fun showResult(result: ScreenResult<String>) {
-        val message = if (result.isSuccess) result.data else "No result"
+    /**
+     * Показ снека с результирующим текстом, либо показ диалога с сообщением о том,
+     * что результата нет, и предложением открыть экран с вводом текста заново.
+     */
+    private fun showResultMap(result: ScreenResult<String>): Observable<MainEvent> {
+        val hasResultData = result.isSuccess && !result.data.isNullOrEmpty()
+        return if (hasResultData) {
+            messageController.show(result.data).skip()
+        } else {
+            Navigation().open(CustomAlertDialogRoute<MainEvent>(
+                    title = "No result",
+                    message = "Try again?",
+                    positiveButtonEvent = Navigation().open(InputFormActivityRoute()))
+            ).toObservable()
+        }
+    }
+
+    /**
+     * Простой показ снека с результирующим текстом
+     */
+    private fun showResultReact(result: ScreenResult<String>) {
+        val hasResultData = result.isSuccess && !result.data.isNullOrEmpty()
+        val message = if (hasResultData) result.data else "No result"
         messageController.show(message)
     }
 }