+++ core-ui/RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 ##### Core-ui
 * ANDDEP-466 Fixed crash when calling `FragmentNavigaotr.popStack` with `popDepth` < 0
 * Fix logging screen names
+* Add additional Route Extra parameters
 ## 0.4.0
 ##### Core-ui
 * Added the ability to set text for the buttons of the standard dialog for switching to settings.